requirements.txt
alabaster==0.7.12
asgiref==3.5.2
Babel==2.8.0
bleach==3.2.1
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.4
Django==4.0.5
docutils==0.16
greenlet==1.1.2
idna==2.10
imagesize==1.2.0
Jinja2==2.11.2
keyring==21.4.0
MarkupSafe==1.1.1
mysql-connector-python==8.0.29
numpy==1.22.4
packaging==20.4
pandas==1.4.2
pkginfo==1.6.0
Pygments==2.7.2
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2020.1
pywin32-ctypes==0.2.0
readme-renderer==28.0
requests==2.24.0
requests-toolbelt==0.9.1
rfc3986==1.4.0
six==1.15.0
snowballstemmer==2.0.0
Sphinx==3.2.1
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
SQLAlchemy==1.4.37
sqlparse==0.4.2
tqdm==4.50.2
twine==3.2.0
tzdata==2022.1
urllib3==1.25.11
webencodings==0.5.1

django-filter~=21.1